Fix gap on right side of page
Using `border-box` box sizing prevents accidentally creating elements
that are wider than the viewport when their width is combined with
padding and margins. Hiding the overflow of the `html` and `body`
elements gets rid of the ~5px gap on the right side of the page that
could be scrolled over to.